requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.0.0
elevenlabs>=1.0.0
python-multipart>=0.0.6
faiss-cpu>=1.7.0
tiktoken>=0.5.0
